move binary tests to integration
